+++ common/src/main/java/com/devnemo/nemos/campfires/world/item/NemosCampfiresItems.java
@@ -54,6 +54,17 @@ public class NemosCampfiresItems {
     public static final Supplier<Item> OAK_REDSTONE_CAMPFIRE = registerCampfireBlockItem("oak_redstone_campfire", NemosCampfiresBlocks.OAK_REDSTONE_CAMPFIRE);
     public static final Supplier<Item> SPRUCE_REDSTONE_CAMPFIRE = registerCampfireBlockItem("spruce_redstone_campfire", NemosCampfiresBlocks.SPRUCE_REDSTONE_CAMPFIRE);
     public static final Supplier<Item> WARPED_REDSTONE_CAMPFIRE = registerCampfireBlockItem("warped_redstone_campfire", NemosCampfiresBlocks.WARPED_REDSTONE_CAMPFIRE);
+    public static final Supplier<Item> ACACIA_LAPIS_CAMPFIRE = registerCampfireBlockItem("acacia_lapis_campfire", NemosCampfiresBlocks.ACACIA_LAPIS_CAMPFIRE);
+    public static final Supplier<Item> BIRCH_LAPIS_CAMPFIRE = registerCampfireBlockItem("birch_lapis_campfire", NemosCampfiresBlocks.BIRCH_LAPIS_CAMPFIRE);
+    public static final Supplier<Item> CHERRY_LAPIS_CAMPFIRE = registerCampfireBlockItem("cherry_lapis_campfire", NemosCampfiresBlocks.CHERRY_LAPIS_CAMPFIRE);
+    public static final Supplier<Item> PALE_OAK_LAPIS_CAMPFIRE = registerCampfireBlockItem("pale_oak_lapis_campfire", NemosCampfiresBlocks.PALE_OAK_LAPIS_CAMPFIRE);
+    public static final Supplier<Item> CRIMSON_LAPIS_CAMPFIRE = registerCampfireBlockItem("crimson_lapis_campfire", NemosCampfiresBlocks.CRIMSON_LAPIS_CAMPFIRE);
+    public static final Supplier<Item> DARK_OAK_LAPIS_CAMPFIRE = registerCampfireBlockItem("dark_oak_lapis_campfire", NemosCampfiresBlocks.DARK_OAK_LAPIS_CAMPFIRE);
+    public static final Supplier<Item> JUNGLE_LAPIS_CAMPFIRE = registerCampfireBlockItem("jungle_lapis_campfire", NemosCampfiresBlocks.JUNGLE_LAPIS_CAMPFIRE);
+    public static final Supplier<Item> MANGROVE_LAPIS_CAMPFIRE = registerCampfireBlockItem("mangrove_lapis_campfire", NemosCampfiresBlocks.MANGROVE_LAPIS_CAMPFIRE);
+    public static final Supplier<Item> OAK_LAPIS_CAMPFIRE = registerCampfireBlockItem("oak_lapis_campfire", NemosCampfiresBlocks.OAK_LAPIS_CAMPFIRE);
+    public static final Supplier<Item> SPRUCE_LAPIS_CAMPFIRE = registerCampfireBlockItem("spruce_lapis_campfire", NemosCampfiresBlocks.SPRUCE_LAPIS_CAMPFIRE);
+    public static final Supplier<Item> WARPED_LAPIS_CAMPFIRE = registerCampfireBlockItem("warped_lapis_campfire", NemosCampfiresBlocks.WARPED_LAPIS_CAMPFIRE);
 
     public static void init() {}