+++ common/src/main/java/com/devnemo/nemos/campfires/world/level/block/NemosCampfiresBlocks.java
@@ -61,6 +61,17 @@ public class NemosCampfiresBlocks {
     public static final Supplier<Block> OAK_REDSTONE_CAMPFIRE = register("oak_redstone_campfire", properties -> new CampfireBlock(false, 1, properties), createCampfireProperties(OAK_LOG.defaultMapColor(), 15));
     public static final Supplier<Block> SPRUCE_REDSTONE_CAMPFIRE = register("spruce_redstone_campfire", properties -> new CampfireBlock(false, 1, properties), createCampfireProperties(SPRUCE_LOG.defaultMapColor(), 15));
     public static final Supplier<Block> WARPED_REDSTONE_CAMPFIRE = register("warped_redstone_campfire", properties -> new CampfireBlock(false, 1, properties), createCampfireProperties(WARPED_STEM.defaultMapColor(), 15));
+    public static final Supplier<Block> ACACIA_LAPIS_CAMPFIRE = register("acacia_lapis_campfire", properties -> new CampfireBlock(false, 1, properties), createCampfireProperties(ACACIA_LOG.defaultMapColor(), 15));
+    public static final Supplier<Block> BIRCH_LAPIS_CAMPFIRE = register("birch_lapis_campfire", properties -> new CampfireBlock(false, 1, properties), createCampfireProperties(BIRCH_LOG.defaultMapColor(), 15));
+    public static final Supplier<Block> CHERRY_LAPIS_CAMPFIRE = register("cherry_lapis_campfire", properties -> new CampfireBlock(false, 1, properties), createCampfireProperties(CHERRY_LOG.defaultMapColor(), 15));
+    public static final Supplier<Block> PALE_OAK_LAPIS_CAMPFIRE = register("pale_oak_lapis_campfire", properties -> new CampfireBlock(false, 1, properties), createCampfireProperties(PALE_OAK_LOG.defaultMapColor(), 15));
+    public static final Supplier<Block> CRIMSON_LAPIS_CAMPFIRE = register("crimson_lapis_campfire", properties -> new CampfireBlock(false, 1, properties), createCampfireProperties(CRIMSON_STEM.defaultMapColor(), 15));
+    public static final Supplier<Block> DARK_OAK_LAPIS_CAMPFIRE = register("dark_oak_lapis_campfire", properties -> new CampfireBlock(false, 1, properties), createCampfireProperties(DARK_OAK_LOG.defaultMapColor(), 15));
+    public static final Supplier<Block> JUNGLE_LAPIS_CAMPFIRE = register("jungle_lapis_campfire", properties -> new CampfireBlock(false, 1, properties), createCampfireProperties(JUNGLE_LOG.defaultMapColor(), 15));
+    public static final Supplier<Block> MANGROVE_LAPIS_CAMPFIRE = register("mangrove_lapis_campfire", properties -> new CampfireBlock(false, 1, properties), createCampfireProperties(MANGROVE_LOG.defaultMapColor(), 15));
+    public static final Supplier<Block> OAK_LAPIS_CAMPFIRE = register("oak_lapis_campfire", properties -> new CampfireBlock(false, 1, properties), createCampfireProperties(OAK_LOG.defaultMapColor(), 15));
+    public static final Supplier<Block> SPRUCE_LAPIS_CAMPFIRE = register("spruce_lapis_campfire", properties -> new CampfireBlock(false, 1, properties), createCampfireProperties(SPRUCE_LOG.defaultMapColor(), 15));
+    public static final Supplier<Block> WARPED_LAPIS_CAMPFIRE = register("warped_lapis_campfire", properties -> new CampfireBlock(false, 1, properties), createCampfireProperties(WARPED_STEM.defaultMapColor(), 15));
 
     private static BlockBehaviour.Properties createCampfireProperties(MapColor mapColor, int lightValue) {
         return BlockBehaviour.Properties.of()